requirements.txt
snowflake-connector-python>=3.6.0
pandas>=2.1.0
numpy>=1.26.0
pytest>=8.0.0
ruff>=0.4.0
mypy>=1.9.0
pandas-stubs>=2.1.0
